Remoção da estrutura antiga AgenciaViagens após renomeação para GoDecola
