--- AgenciaViagens/AgenciaViagens.API/AgenciaViagens.API.http
@@ -1,6 +1,0 @@
-@AgenciaViagens.API_HostAddress = http://localhost:5126
-
-GET {{AgenciaViagens.API_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
