+++ GoDecola/GoDecola.API/GoDecola.API.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AgenciaViagens.Infrastructure\AgenciaViagens.Infrastructure.csproj" />
+    <ProjectReference Include="..\GoDecola.Infrastructure\GoDecola.Infrastructure.csproj" />
   </ItemGroup>
 
   <ItemGroup>
