+++ GoDecola/GoDecola.Domain/Entities/PackageMedia.cs
@@ -4,17 +4,18 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AgenciaViagens.Domain.Entities
+namespace GoDecola.Domain.Entities
 {
 
-    public class WishList
+    public class PackageMedia
     {
         public Guid Id { get; set; }
-        public Guid UserId { get; set; }
-        public User? User { get; set; }
-
         public Guid PackageId { get; set; }
         public TravelPackage? Package { get; set; }
+
+        public string MediaType { get; set; } = null!;
+        public string FileName { get; set; } = null!;
+        public string Url { get; set; } = null!;
     }
 
 }
