+++ GoDecola/GoDecola.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.14.36221.1 d17.14
+VisualStudioVersion = 17.14.36221.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AgenciaViagens.API", "AgenciaViagens.API\AgenciaViagens.API.csproj", "{DD02A661-3819-4D9B-A1BA-30699F2C0C18}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoDecola.API", "GoDecola.API\GoDecola.API.csproj", "{DD02A661-3819-4D9B-A1BA-30699F2C0C18}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AgenciaViagens.Domain", "AgenciaViagens.Domain\AgenciaViagens.Domain.csproj", "{B75F9CA9-DED2-45FF-AE03-8E44C4C62966}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoDecola.Domain", "GoDecola.Domain\GoDecola.Domain.csproj", "{B75F9CA9-DED2-45FF-AE03-8E44C4C62966}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AgenciaViagens.Infrastructure", "AgenciaViagens.Infrastructure\AgenciaViagens.Infrastructure.csproj", "{79CA8C7B-A697-4CF4-B949-6BFCF3E7EDD4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoDecola.Infrastructure", "GoDecola.Infrastructure\GoDecola.Infrastructure.csproj", "{79CA8C7B-A697-4CF4-B949-6BFCF3E7EDD4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
